src/mcqgenerator/MCQgenerator.py
import os
import json
import pandas as pd
import traceback
from dotenv import load_dotenv
from src.mcqgenerator.utils import read_file,get_table_data
from src.mcqgenerator.logger import logging

# Import necessary classes from langchain
from langchain_google_genai import ChatGoogleGenerativeAI
from langchain.prompts import PromptTemplate
from langchain.chains import LLMChain
from langchain.chains import SequentialChain

# Load environment variables from the .env file
load_dotenv()

# Get the Google API key from environment variables
KEY = os.getenv("GOOGLE_API_KEY")

# Initialize the Gemini model (using latest available model)
llm = ChatGoogleGenerativeAI(
    google_api_key=KEY,
    model="gemini-pro-latest",  # Updated to use latest model
    temperature=0.5
)

# Create the prompt template for quiz generation.
TEMPLATE = """
Text:{text}
You are an expert MCQ maker. Given the above text, it is your job to \
create a quiz of {number} multiple choice questions for {subject} students in {tone} tone. 
Make sure the questions are not repeated and check all the questions to be conforming the text as well.

IMPORTANT: Your response must be ONLY valid JSON in the exact format shown below. Do not include any explanatory text, markdown formatting, or code blocks. Just return the raw JSON.

### RESPONSE_JSON
{response_json}

Return ONLY the JSON object with {number} MCQs, nothing else.
"""

quiz_generation_prompt = PromptTemplate(
    input_variables=["text", "number", "subject", "tone", "response_json"],
    template=TEMPLATE
)

# Create the LLMChain for generating the quiz.
quiz_chain = LLMChain(llm=llm, prompt=quiz_generation_prompt, output_key="quiz", verbose=True)

# Create the prompt template for evaluating the generated quiz.
TEMPLATE2 = """
You are an expert english grammarian and writer. Given a Multiple Choice Quiz for {subject} students.\
You need to evaluate the complexity of the question and give a complete analysis of the quiz. Only use at max 50 words for complexity analysis. 
if the quiz is not at per with the cognitive and analytical abilities of the students,\
update the quiz questions which needs to be changed and change the tone such that it perfectly fits the student abilities
Quiz_MCQs:
{quiz}

Check from an expert English Writer of the above quiz:
"""

quiz_evaluation_prompt = PromptTemplate(input_variables=["subject", "quiz"], template=TEMPLATE2)

# Create the LLMChain for reviewing the quiz.
review_chain = LLMChain(llm=llm, prompt=quiz_evaluation_prompt, output_key="review", verbose=True)

# Create the final SequentialChain that runs the quiz and review chains in order.
generate_evaluate_chain = SequentialChain(
    chains=[quiz_chain, review_chain],
    input_variables=["text", "number", "subject", "tone", "response_json"],
    output_variables=["quiz", "review"],
    verbose=True
)