requirements.txt
# Core LLM and LangChain (pinned for compatibility)
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.12
langchain-google-genai==1.0.7
google-generativeai==0.7.2

# Web UI
streamlit==1.37.1

# Env variables(key value pair)
python-dotenv==1.0.1

# File handling
PyPDF2==3.0.1
pdfminer.six==20240706

# Pydantic pinned due to upstream compatibility
pydantic==2.8.2


-e .
